Make constant memory opt-in, spill large statics to global memory (#217)
* Allow address spaces to propagate to LLVM

This looks like it was code that wasn't deleted after the refactor in
https://github.com/Rust-GPU/Rust-CUDA/commit/decda8784d10fc4e4a9b9e385e2a4322cfd07d79

* Spill large statics from constant to global memory

This isn't fully correct, as ideally we keep track
of what we have put into constant memory and when
it is filled up spill instdead of only spilling when
a static is big. But, this is materially better than
what is there (which is a runtime error).

An argument can be made to just _always_ use global
memory and we don't have to worry about getting the
packing right.

Fixes https://github.com/Rust-GPU/Rust-CUDA/issues/208.

See also the debugging and discussion in
https://github.com/Rust-GPU/Rust-CUDA/pull/216

* Add `--use-constant-memory-space` flag, off by default

* Make it clear that `#[cuda_std::address_space(constant)]` still works